publish samples
