+++ sdk/cognitivelanguage/ai-language-text/samples/v1/javascript/batching.js
