+++ sdk/cognitivelanguage/ai-language-text/samples/v1/javascript/healthcare.js
@@ -6,19 +6,13 @@
  * healthcare-related entities in some documents and prints them to the
  * console.
  *
- * @summary detects healthcare entities in a piece of text and creates an FHIR representation
+ * @summary detects healthcare entities in a piece of text
  */
 
-import {
-  AnalyzeBatchAction,
-  AzureKeyCredential,
-  KnownFhirVersion,
-  TextAnalysisClient,
-} from "@azure/ai-language-text";
+const { AzureKeyCredential, TextAnalysisClient } = require("@azure/ai-language-text");
 
 // Load the .env file if it exists
-import * as dotenv from "dotenv";
-dotenv.config();
+require("dotenv").config();
 
 // You will need to set these environment variables or edit the following values
 const endpoint = process.env["ENDPOINT"] || "<cognitive language service endpoint>";
@@ -30,14 +24,13 @@ const documents = [
   "Patient does not suffer from high blood pressure.",
 ];
 
-export async function main() {
+async function main() {
   console.log("== Healthcare Sample ==");
 
   const client = new TextAnalysisClient(endpoint, new AzureKeyCredential(apiKey));
-  const actions: AnalyzeBatchAction[] = [
+  const actions = [
     {
       kind: "Healthcare",
-      fhirVersion: KnownFhirVersion["4.0.1"],
     },
   ];
   const poller = await client.beginAnalyzeBatch(actions, documents, "en");
@@ -87,9 +80,6 @@ export async function main() {
           }
         }
       }
-      if (result.fhirBundle) {
-        console.log(`FHIR object: ${JSON.stringify(result.fhirBundle, undefined, 2)}`);
-      }
     }
   }
 }
@@ -97,3 +87,5 @@ export async function main() {
 main().catch((err) => {
   console.error("The sample encountered an error:", err);
 });
+
+module.exports = { main };
