+++ sdk/cognitivelanguage/ai-language-text/samples/v1/javascript/package.json
@@ -1,14 +1,10 @@
 {
-  "name": "@azure-samples/ai-language-text-ts-beta",
+  "name": "@azure-samples/ai-language-text-js",
   "private": true,
   "version": "1.0.0",
-  "description": "Azure Cognitive Language Service client library samples for TypeScript (Beta)",
+  "description": "Azure Cognitive Language Service client library samples for JavaScript",
   "engines": {
     "node": ">=12.0.0"
-  },
-  "scripts": {
-    "build": "tsc",
-    "prebuild": "rimraf dist/"
   },
   "repository": {
     "type": "git",
@@ -30,13 +26,8 @@
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/cognitivelanguage/ai-language-text",
   "dependencies": {
-    "@azure/ai-language-text": "next",
+    "@azure/ai-language-text": "latest",
     "dotenv": "latest",
     "@azure/identity": "^2.0.1"
-  },
-  "devDependencies": {
-    "@types/node": "^12.0.0",
-    "typescript": "~4.6.0",
-    "rimraf": "latest"
   }
 }
