+++ sdk/cognitivelanguage/ai-language-text/samples/v1/typescript/sample.env
@@ -1,7 +1,7 @@
-# Used to authenticate with the Azure Cognitive Language Service. Retrieve these
-# values from a Azure Cognitive Language Service instance in the Azure Portal.
+# Used to authenticate with the Azure Cognitive Service for Language. Retrieve these
+# values from a Cognitive Services instance in the Azure Portal.
 ENDPOINT="https://<resource name>.cognitiveservies.azure.com/"
-LANGUAGE_API_KEY="<cognitive language resource api key>"
+LANGUAGE_API_KEY="<text analytics api key>"
 
 # Used to authenticate using Azure AD as a service principal for role-based
 # authentication.
