+++ sdk/cognitivelanguage/ai-language-text/samples/v1/typescript/src/modelVersion.ts
