+++ sdk/cognitivelanguage/ai-language-text/samples/v1/typescript/src/opinionMining.ts
