+++ sdk/cognitivelanguage/ai-language-text/samples/v1/typescript/tsconfig.json
